Fix Windows installation error
Described in Issue #1. Switched from floats to doubles to calculate the
matrix of permutation enrichment scores. This has the effect of
increasing the runtime of fast_ssgsea by around 40%, so all .rds files
and figures had to be updated. The README now includes detailed
instructions for users who wish to use the improved version of the code.